+++ src/test/converter2/issues/gh2165.ts
@@ -1,0 +1,6 @@
+/**
+ * 'module' comment
+ * @module
+ */
+
+var globalFile = true;